Moves RamSlotComposer to runtime/testing
